We should use near_network_primitives directly
