+++ chain/client/src/sync.rs
@@ -13,10 +13,7 @@ use rand::seq::{IteratorRandom, SliceRandom};
 use rand::{thread_rng, Rng};
 
 use near_chain::{Chain, RuntimeAdapter};
-use near_network::types::{
-    AccountOrPeerIdOrHash, NetworkResponses, PeerManagerMessageRequest, ReasonForBan,
-};
-use near_network::{FullPeerInfo, NetworkRequests, PeerManagerAdapter};
+use near_network::{FullPeerInfo, NetworkRequests, NetworkResponses, PeerManagerAdapter};
 use near_primitives::block::Tip;
 use near_primitives::hash::CryptoHash;
 use near_primitives::network::PeerId;
@@ -33,6 +30,8 @@ use near_chain::chain::{ApplyStatePartsRequest, StateSplitRequest};
 use near_client_primitives::types::{
     DownloadStatus, ShardSyncDownload, ShardSyncStatus, SyncStatus,
 };
+use near_network::types::PeerManagerMessageRequest;
+use near_network_primitives::types::AccountOrPeerIdOrHash;
 use near_primitives::shard_layout::ShardUId;
 
 /// Maximum number of block headers send over the network.
@@ -277,7 +276,7 @@ impl HeaderSync {
                                         PeerManagerMessageRequest::NetworkRequests(
                                             NetworkRequests::BanPeer {
                                                 peer_id: peer.peer_info.id.clone(),
-                                                ban_reason: ReasonForBan::HeightFraud,
+                                                ban_reason: near_network_primitives::types::ReasonForBan::HeightFraud,
                                             },
                                         ),
                                     );
@@ -1285,14 +1284,13 @@ mod test {
     use near_chain::{ChainGenesis, Provenance};
     use near_crypto::{KeyType, PublicKey};
     use near_network::test_utils::MockPeerManagerAdapter;
-    use near_network::types::PeerChainInfoV2;
-    use near_network::PeerInfo;
     use near_primitives::block::{Approval, Block, GenesisId};
     use near_primitives::network::PeerId;
 
     use super::*;
     use crate::test_utils::TestEnv;
     use near_network::routing::routing::EdgeInfo;
+    use near_network_primitives::types::PeerInfo;
     use near_primitives::merkle::PartialMerkleTree;
     use near_primitives::types::EpochId;
     use near_primitives::validator_signer::InMemoryValidatorSigner;
@@ -1349,7 +1347,7 @@ mod test {
         let mut sync_status = SyncStatus::NoSync;
         let peer1 = FullPeerInfo {
             peer_info: PeerInfo::random(),
-            chain_info: PeerChainInfoV2 {
+            chain_info: near_network_primitives::types::PeerChainInfoV2 {
                 genesis_id: GenesisId {
                     chain_id: "unittest".to_string(),
                     hash: *chain.genesis().hash(),
