+++ chain/client/tests/bug_repros.rs
@@ -17,7 +17,8 @@ use near_crypto::{InMemorySigner, KeyType};
 use near_logger_utils::init_test_logger;
 use near_network::types::NetworkRequests::PartialEncodedChunkMessage;
 use near_network::types::{PeerManagerMessageRequest, PeerManagerMessageResponse};
-use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses, PeerInfo};
+use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses};
+use near_network_primitives::types::PeerInfo;
 use near_primitives::block::Block;
 use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::AccountId;
