+++ chain/client/tests/consensus.rs
@@ -13,7 +13,8 @@ mod tests {
     use near_client::{ClientActor, ViewClientActor};
     use near_logger_utils::init_integration_logger;
     use near_network::types::PeerManagerMessageRequest;
-    use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses, PeerInfo};
+    use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses};
+    use near_network_primitives::types::PeerInfo;
     use near_primitives::block::{Approval, ApprovalInner};
     use near_primitives::types::{AccountId, BlockHeight};
