+++ chain/jsonrpc/Cargo.toml
@@ -28,6 +28,7 @@ near-jsonrpc-client = { path = "client" }
 near-jsonrpc-primitives = { path = "../jsonrpc-primitives" }
 near-jsonrpc-adversarial-primitives = { path = "../jsonrpc-adversarial-primitives", optional = true }
 near-rpc-error-macro = { path = "../../tools/rpctypegen/macro" }
+near-network-primitives = { path = "../../chain/network-primitives" }
 
 [features]
 dump_errors_schema = ["near-rpc-error-macro/dump_errors_schema"]
