+++ chain/jsonrpc/src/lib.rs
@@ -40,22 +40,26 @@ use near_metrics::{Encoder, TextEncoder};
 #[cfg(feature = "test_features")]
 use near_network::routing::routing::GetRoutingTableResult;
 #[cfg(feature = "test_features")]
-use near_network::types::{
-    GetPeerId, NetworkAdversarialMessage, NetworkViewClientMessages, PeerManagerMessageRequest,
-    SetAdvOptions,
-};
+use near_network::types::PeerManagerMessageRequest;
 #[cfg(feature = "sandbox")]
-use near_network::types::{NetworkSandboxMessage, SandboxResponse};
+use near_network::types::SandboxResponse;
+#[cfg(feature = "test_features")]
+use near_network::types::SetAdvOptions;
 #[cfg(all(
     feature = "test_features",
     feature = "protocol_feature_routing_exchange_algorithm"
 ))]
 use near_network::types::{SetRoutingTable, StartRoutingTableSync};
-use near_network::{NetworkClientMessages, NetworkClientResponses};
 #[cfg(feature = "test_features")]
 use near_network::{
-    PeerManagerActor, RoutingTableActor, RoutingTableMessages, RoutingTableMessagesResponse,
+    GetPeerId, PeerManagerActor, RoutingTableActor, RoutingTableMessages,
+    RoutingTableMessagesResponse,
 };
+use near_network::{NetworkClientMessages, NetworkClientResponses};
+#[cfg(feature = "sandbox")]
+use near_network_primitives::types::NetworkSandboxMessage;
+#[cfg(feature = "test_features")]
+use near_network_primitives::types::{NetworkAdversarialMessage, NetworkViewClientMessages};
 use near_primitives::hash::CryptoHash;
 use near_primitives::serialize::BaseEncode;
 use near_primitives::transaction::SignedTransaction;
