+++ chain/network-primitives/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod types;
+pub mod utils;
