+++ genesis-tools/genesis-csv-to-json/Cargo.toml
@@ -17,6 +17,7 @@ near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
 node-runtime = { path = "../../runtime/runtime" }
 near-network = { path = "../../chain/network" }
+near-network-primitives = { path = "../../chain/network-primitives" }
 
 [dev-dependencies]
 tempfile = "3"
