+++ genesis-tools/genesis-csv-to-json/src/csv_parser.rs
@@ -9,7 +9,7 @@ use near_primitives::time::Utc;
 use serde::{Deserialize, Serialize};
 
 use near_crypto::{KeyType, PublicKey};
-use near_network::PeerInfo;
+use near_network_primitives::types::PeerInfo;
 use near_primitives::account::{AccessKey, AccessKeyPermission, Account, FunctionCallPermission};
 use near_primitives::hash::{hash, CryptoHash};
 use near_primitives::receipt::{ActionReceipt, Receipt, ReceiptEnum};
