+++ integration-tests/Cargo.toml
@@ -47,6 +47,7 @@ near-vm-runner = { path = "../runtime/near-vm-runner" }
 nearcore = { path = "../nearcore" }
 node-runtime = { path = "../runtime/runtime" }
 testlib = { path = "../test-utils/testlib" }
+near-network-primitives = { path = "../chain/network-primitives" }
 
 [features]
 performance_stats = ["nearcore/performance_stats", "near-network/performance_stats"]
