+++ integration-tests/tests/client/chunks_management.rs
@@ -16,8 +16,9 @@ use near_chunks::{
 use near_client::test_utils::setup_mock_all_validators;
 use near_client::{ClientActor, GetBlock, ViewClientActor};
 use near_logger_utils::init_test_logger;
-use near_network::types::{AccountIdOrPeerTrackingShard, PeerManagerMessageRequest};
-use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses, PeerInfo};
+use near_network::types::PeerManagerMessageRequest;
+use near_network::{NetworkClientMessages, NetworkRequests, NetworkResponses};
+use near_network_primitives::types::{AccountIdOrPeerTrackingShard, PeerInfo};
 use near_primitives::hash::CryptoHash;
 use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::AccountId;
