+++ nearcore/Cargo.toml
@@ -44,6 +44,7 @@ near-telemetry = { path = "../chain/telemetry" }
 near-epoch-manager = { path = "../chain/epoch_manager" }
 near-performance-metrics = { path = "../utils/near-performance-metrics" }
 near-vm-runner = { path = "../runtime/near-vm-runner"}
+near-network-primitives = { path = "../chain/network-primitives" }
 
 delay-detector = { path = "../tools/delay_detector", optional = true }
